add some more test
